remove unused classes
