+++ src/NServiceBus.AcceptanceTesting/NServiceBus.AcceptanceTesting.csproj
@@ -68,7 +68,6 @@
     <Compile Include="Support\EndpointRunner.cs" />
     <Compile Include="Support\FailTestOnErrorMessageFeature.cs" />
     <Compile Include="Support\IConfigureEndpointTestExecution.cs" />
-    <Compile Include="Support\IConfigureSupportedScenariosForTestExecution.cs" />
     <Compile Include="Support\IEndpointSetupTemplate.cs" />
     <Compile Include="Support\IScenarioWithEndpointBehavior.cs" />
     <Compile Include="Support\IWhenDefinition.cs" />
